Binary Search Tree : Lowest Common Ancestor
Function Description

Complete the function lca in the editor below. It should return a pointer to the lowest common ancestor node of the two values given.

lca has the following parameters: 
- root: a pointer to the root node of a binary search tree 
- v1: a node.data value 
- v2: a node.data value

Input Format

The first line contains an integer, , the number of nodes in the tree. 
The second line contains  space-separated integers representing  values. 
The third line contains two space-separated integers,  and .

To use the test data, you will have to create the binary search tree yourself. Here on the platform, the tree will be created for you.