Create Delete_a_Node_Singly_Linked_list_given_head.java
You’re given the pointer to the head node of a linked list and the position of a node to delete. Delete the node at the given position and return the head node. A position of 0 indicates head, a position of 1 indicates one node away from the head and so on. The list may become empty after you delete the node.

Input Format

You have to complete the deleteNode(SinglyLinkedListNode* llist, int position) method which takes two arguments - the head of the linked list and the position of the node to delete. You should NOT read any input from stdin/console. position will always be at least 0 and less than the number of the elements in the list.

The first line of input contains an integer , denoting the number of elements in the linked list. 
The next  lines contain an integer each in a new line, denoting the elements of the linked list in the order. 
The last line contains an integer  denoting the position of the node that has to be deleted form the linked list.

Input - 8
20
6
2
19
7
4
15
9
3